requirement.txt
anyio==3.5.0
asgiref==3.4.1
click==8.0.3
fastapi==0.72.0
h11==0.13.0
idna==3.3
pydantic==1.9.0
PyMySQL==1.0.2
python-ESL==1.4.18
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.0.1
uvicorn==0.17.0
